test: Youtube Player를 Global Template에 위치
